test: user 테스트 실패 수정
